url/host-parser.html

<!doctype html>
<meta charset=utf-8>
<title>Testing the host parser</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
let setup = async_test("Loading data…")
setup.step(() => {
  const tests = { "asciihost": null, "IdnaTest": null }
  let testNumber = Object.keys(tests).length
  Object.keys(tests).forEach((val) => {
    let request = new XMLHttpRequest()
    request.open("GET", val + ".json")
    request.send()
    request.responseType = "json"
    request.onload = setup.step_func(() => {
      tests[val] = request.response
      testNumber--
      if(testNumber === 0) {
        Object.keys(tests).forEach((val) => {
          runHostTests(val, tests[val])
        })
        setup.done()
      }
    })
  })
})

function makeURL(type, input) {
  input = "https://" + input + "/x"
  if(type === "url") {
    return new URL(input)
  } else {
    const url = document.createElement(type)
    url.href = input
    return url
  }
}

function runHostTests(source, hostTests) {
  for(var i = 0, l = hostTests.length; i < l; i++) {
    let hostTest = hostTests[i]
    if (typeof hostTest === "string") {
      continue // skip comments
    }
    let outputInput = hostTest.input
    if(hostTest.input.length < 4) {
      outputInput = encodeURI(hostTest.input)
    }

    const typeName = { "url": "URL", "a": "<a>", "area": "<area>" }
    ;["url", "a", "area"].forEach((type) => {
      if(!hostTest.onlySetters) {
        test(() => {
          if(hostTest.output !== null) {
            const url = makeURL("url", hostTest.input)
            assert_equals(url.host, hostTest.output)
            assert_equals(url.hostname, hostTest.output)
            assert_equals(url.pathname, "/x")
            assert_equals(url.href, "https://" + hostTest.output + "/x")
          } else {
            if(type === "url") {
              assert_throws(new TypeError, () => makeURL("url", hostTest.input))
            } else {
              const url = makeURL(type, hostTest.input)
              assert_equals(url.host, "")
              assert_equals(url.hostname, "")
              assert_equals(url.pathname, "")
              assert_equals(url.href, "https://" + hostTest.input + "/x")
            }
          }
        }, "Parse host using " + typeName[type] + ": " + outputInput + " (from: " + source + ")")
      }

      ;["host", "hostname"].forEach((val) => {
        test(() => {
          const url = makeURL(type, "x")
          url[val] = hostTest.input
          if(hostTest.output !== null) {
            assert_equals(url[val], hostTest.output)
          } else {
            assert_equals(url[val], "x")
          }
        }, "Parse host using " + typeName[type] + "." + val + ": " + outputInput + " (from: " + source + ")")
      })
    })
  }
}
</script>
